refactor: move generic ros2 tools to generic folder
