+++ src/rai_core/rai/tools/ros2/generic/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2024 Robotec.AI
+# Copyright (C) 2025 Robotec.AI
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,38 +12,36 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .generic import (
-    CallROS2ServiceTool,
+from .actions import (
     CancelROS2ActionTool,
     GetROS2ActionsNamesAndTypesTool,
+    ROS2ActionToolkit,
+    StartROS2ActionTool,
+)
+from .services import (
+    CallROS2ServiceTool,
+    GetROS2ServicesNamesAndTypesTool,
+    ROS2ServicesToolkit,
+)
+from .toolkit import ROS2Toolkit
+from .topics import (
     GetROS2ImageTool,
     GetROS2MessageInterfaceTool,
-    GetROS2ServicesNamesAndTypesTool,
     GetROS2TopicsNamesAndTypesTool,
     GetROS2TransformTool,
     PublishROS2MessageTool,
     ReceiveROS2MessageTool,
-    ROS2ActionToolkit,
-    ROS2ServicesToolkit,
-    ROS2Toolkit,
     ROS2TopicsToolkit,
-    StartROS2ActionTool,
-)
-from .simple import (
-    GetROS2ImageConfiguredTool,
-    GetROS2TransformConfiguredTool,
 )
 
 __all__ = [
     "CallROS2ServiceTool",
     "CancelROS2ActionTool",
     "GetROS2ActionsNamesAndTypesTool",
-    "GetROS2ImageConfiguredTool",
     "GetROS2ImageTool",
     "GetROS2MessageInterfaceTool",
     "GetROS2ServicesNamesAndTypesTool",
     "GetROS2TopicsNamesAndTypesTool",
-    "GetROS2TransformConfiguredTool",
     "GetROS2TransformTool",
     "PublishROS2MessageTool",
     "ROS2ActionToolkit",
